# docs/swagger.yaml
definitions:
  httpError.HTTPError:
    properties:
      message:
        type: string
    type: object
...
  usersRoute.Pair:
    properties:
      group_id:
        type: string
      mentor_id:
        type: string
    type: object
  usersRoute.reqCreateHelp:
    properties:
      goal:
        type: string
      requests:
        items:
          $ref: '#/definitions/usersRoute.Pair'
        type: array
    required:
    - goal
    - requests
    type: object
...
  usersRoute.resGetProfile:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      name:
        type: string
    type: object
...
  usersRoute.respUploadAvatarDto:
    properties:
      url:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /api/group/requests:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/usersRoute.reqCreateHelp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Кинуть запрос ментору
      tags:
      - Users
...
  /api/user/profile:
    get:
      consumes:
      - application/json
      description: Авторизация юзера
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usersRoute.resGetProfile'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/httpError.HTTPError'
      summary: Получить инфу о себе
      tags:
      - Users
  /api/user/uploadAvatar:
    post:
      consumes:
      - application/json
      description: Загрузка аватарки. Возвращает ссылку на аватарку, которая действует
        7 дней
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usersRoute.respUploadAvatarDto'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/httpError.HTTPError'
      summary: загрузка аватарки
      tags:
      - Users
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"